# SchoolStaff/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SchoolStaff, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Circulars": {
      "get": {
        "tags": [
          "Circulars"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Circular"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Circulars"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Circular"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Circular"
            }
          }
        }
      }
    },
    "/api/Circulars/{id}": {
      "get": {
        "tags": [
          "Circulars"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Circular"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Circulars"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Circular"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Circulars"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Circular": {
      "required": [
        "notificationPostedBy"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "notificationDate": {
          "format": "date-time",
          "type": "string"
        },
        "informationText": {
          "type": "string"
        },
        "notificationPostedBy": {
          "type": "string"
        },
        "notificationPostedByNavigation": {
          "$ref": "#/definitions/staff"
        }
      }
    },
...
    "staff": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "staffName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "circulars": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Circular"
          }
        }
      }
    }
  }
}